Create globalVariables.json
This file will be used to share variables across all devices.

reqID_counter will be used to ensure that the id of a request is always unique. Foreseeable problem with this feature: If using a separate device (i.e. a smart screen) a request would be required to obtain the reqID_counter variable. Obviously, this would be problematic as each request from another device would up the counter by two (if it even worked)