Code Refactored
The data.json file was opened multiple times unnecessarily. It has now been changed in main.py so it is just opened at the start of the for loop. Similar changes have been made in reciever.py and HS_Bot.py.

I was planning to move the command functions to functions.py in order to take up less space in main.py but it just caused the same issue in functions.py. I decided to leave them in main.py